requirements.txt
cloudscraper
requests
beautifulsoup4
Pillow